pyproject.toml
[tool.poetry]
name = "eqnn"
version = "0.1.0"
description = "Equivariant Quantum Neural Network Project"
authors = ["Giuseppe D'Ambruoso <g.dambruoso2@phd.poliba.it>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

absl-py = "2.3.1"
antlr4-python3-runtime = "4.9.3"
appdirs = "1.4.4"
asttokens = "3.0.0"
astunparse = "1.6.3"
attrs = "25.4.0"
autograd = "1.8.0"
autoray = "0.8.2"
beautifulsoup4 = "4.14.2"
black = "25.11.0"
bleach = "6.3.0"
build = "1.3.0"
cachetools = "6.2.2"
certifi = "2025.11.12"
charset-normalizer = "3.4.4"
click = "8.3.1"
contourpy = "1.3.3"
cycler = "0.12.1"
decorator = "5.2.1"
defusedxml = "0.7.1"
docopt = "0.6.2"
fastjsonschema = "2.21.2"
filelock = "3.20.0"
flake8 = "7.3.0"
flatbuffers = "25.9.23"
fonttools = "4.60.1"
fsspec = "2025.10.0"
gast = "0.6.0"
google-pasta = "0.2.0"
grpcio = "1.76.0"
h5py = "3.15.1"
hydra-core = "1.3.2"
idna = "3.11"
isort = "7.0.0"
Jinja2 = "3.1.6"
joblib = "1.5.2"
jsonschema = "4.25.1"
keras = "3.12.0"
kiwisolver = "1.4.9"
libclang = "18.1.1"
Markdown = "3.10"
markdown-it-py = "4.0.0"
MarkupSafe = "3.0.3"
matplotlib = "3.10.7"
mistune = "3.1.4"
ml_dtypes = "0.5.3"
mpmath = "1.3.0"
mypy = "1.18.2"
namex = "0.1.0"
nbclient = "0.10.2"
nbconvert = "7.16.6"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
networkx = "3.5"
numpy = "2.2.6"
omegaconf = "2.3.0"
opencv-python = "4.12.0.88"
opt-einsum = "3.4.0"
optree = "0.18.0"
packaging = "25.0"
pennylane = "0.44.0"
pennylane-lightning = "0.44.0"
pillow = "12.0.0"
pipreqs = "0.5.0"
platformdirs = "4.5.0"
protobuf = "6.33.1"
psutil = "7.1.3"
pycodestyle = "2.14.0"
pyflakes = "3.4.0"
Pygments = "2.19.2"
pyparsing = "3.2.5"
python-dateutil = "2.9.0.post0"
pytokens = "0.3.0"
PyYAML = "6.0.3"
referencing = "0.37.0"
requests = "2.32.5"
rich = "14.2.0"
rustworkx = "0.17.1"
setuptools = "80.9.0"
six = "1.17.0"
soupsieve = "2.8"
sympy = "1.14.0"
tensorboard = "2.20.0"
tensorboard-data-server = "0.7.2"
termcolor = "3.2.0"
tinycss2 = "1.4.0"
tomlkit = "0.13.3"
torch = "2.9.1"
torchvision = "0.24.1"
tqdm = "4.67.1"
typing-extensions = "4.15.0"
urllib3 = "2.5.0"
wcwidth = "0.2.14"
Werkzeug = "3.1.3"
wheel = "0.45.1"
wrapt = "2.0.1"
pytest = "^9.0.1"
pytest-cov = "^7.0.0"
pandas = "^2.3.3"
scikit-learn = "^1.7.2"
hydra-joblib-launcher = "^1.2.0"
seaborn = "^0.13.2"
kagglehub = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
flake8 = "^7.3.0"
black = "^25.11.0"
isort = "^7.0.0"
mypy = "^1.18.2"

# ---------------------------------------------------------------------------
# Formatting tools
# ---------------------------------------------------------------------------
ipykernel = "^7.1.0"

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 88
include_trailing_comma = true
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = false
strict_optional = false

# ---------------------------------------------------------------------------
# Build system
# ---------------------------------------------------------------------------

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
